i64 for fungible asset delta, use BigIntSum for sql usage

+++ crates/store/src/db/schema.rs
@@ -13,7 +13,7 @@ diesel::table! {
         account_id -> Binary,
         block_num -> BigInt,
         faucet_id -> Binary,
-        delta -> Integer,
+        delta -> BigInt,
     }
 }
 
@@ -22,7 +22,7 @@ diesel::table! {
         account_id -> Binary,
         block_num -> BigInt,
         vault_key -> Binary,
-        is_remove -> Integer, // TODO why is this not a bool?
+        is_remove -> Integer, // TODO consider migration to Boolean?
     }
 }
